infra/data-generator/requirements.txt
confluent-kafka>=2.3
psycopg2-binary>=2.9
fastavro>=1.9
boto3>=1.34
attrs==25.3.0
orjson==3.11.3
httpx==0.28.1
cachetools==6.2.0
Authlib==1.6.3
